+++ index.html
@@ -4,6 +4,7 @@
   <meta charset="utf-8">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="css/leaflet.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
@@ -231,5 +232,7 @@
   </template>
 
   <script src="/js/main.js" type="module"></script>
+  <script src="/js/leaflet.js.map"></script>
+  <script src ="js/leaflet.js"></script>
 </body>
 </html>
