+++ js/form-processing.js
@@ -115,4 +115,4 @@ const filterChangeHandler = function (evt) {
 
 adRoomNumberElement.addEventListener('change', filterChangeHandler);
 
-export { inactivateForm, activateForm };
+export { inactivateForm, activateForm, adFormElement };
